Move resume components to html route dir
